--- UIFragmentsJorge/app/src/main/java/com/uc3m/foodstuff/ui/login/LoginViewModelFactory.kt
@@ -1,25 +1,0 @@
-package com.uc3m.foodstuff.ui.login
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.uc3m.foodstuff.data.LoginDataSource
-import com.uc3m.foodstuff.data.LoginRepository
-
-/**
- * ViewModel provider factory to instantiate LoginViewModel.
- * Required given LoginViewModel has a non-empty constructor
- */
-class LoginViewModelFactory : ViewModelProvider.Factory {
-
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(LoginViewModel::class.java)) {
-            return LoginViewModel(
-                    loginRepository = LoginRepository(
-                            dataSource = LoginDataSource()
-                    )
-            ) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
